Port to PyQT5, rm goesr_nc_concat/nco dep
A couple of small improvements thrown in including:
 - default layout to 1 col, 2 rows
 - more consice netcdf preview/summary

goesr_nc_concat really seems to be a pain overall, I've always had,
and seen a lot of trouble with installing the pip and conda nco
dependencies.

I've replaced goesr_nc_concat completely with the aggregoes
functionality. It appears to be somewhat faster as well, and
I'm excited that we'll be able to tie in the progress bar to
aggregation to actually be meaniningfull.