Add setup.py and entry point
pyntpg can now be installed using pip and has an entry point so that
`pyntpg` on the command line will bring up the gui.

Unfortunately, with Python 2.7, it is not possible to install PyQt5
as a requirement dependency, so the user will have to have that
installed to be able to run, luckily anaconda comes with PyQt5 by
default.

resolves #2