Fix unresponsive after select color
Perhaps in updating to pyqt5, the color selector and/or associated
mutex no longer work properly.

Fixed. Working more in the way it was intended. Setting the parent
of the QColorDialog window prevents refocus on it while it's open,
so mostly solved by setting parent.

Bonus, color that appears is what was previously selected and
selecting a color and then pressing cancel doesn't produce any
change.

Fixes !8