Synchronization between axes!
Wow, that was actually pretty easy.

There are now three keyboard shorcuts while a plot
open.

 - X -> will synchronize X axes
 - Y -> will synchronize Y axes
 - R -> will redraw plot, because sometimes it just
    gets messed up

This commit also introduces a status bar to the bottom
of the plot tab, however, it's not currenly hooked up
to anything, but will  be used in the future for, eg #14
to alert user about exceptions.

Fixes #10