Adds message connections to status bar
Adds a tree of pyqtSignals that can propagate
error or other status messages up to the plot
tab to be displayed in the status bar.

Some error handling wrapped and send messages
to this, specifically plotting, and any error
with the datetime x-axis widget.

Fixes #14, though there should be ongoing
improvements to this.