Upd for cftime 1.2.0 would o/w break time axis. Bump v0.3
As of cftime 1.2.0, that library no longer returns a native
datetime.datetime object. Instead, it returns a custom object that does
not inherit. So... this a) doesn't trigger detection by isinstance(x,
datetime). b) the custom type is not understood by pyqt5 for the
datetime widget.

Cftime also changed the signature on _dateparse to include the calendar,
so additionally, new logic has been added to maintain backwards
compatibility but also be forward compatible with the two argument
version.